+++ octarine/__init__.py
@@ -2,5 +2,5 @@ from . import utils
 from .viewer import *
 from .__version__ import __version__
 from .conversion import register_converter, get_converter
-from .utils import VoxelCloud
+from .utils import VoxelCloud, VoxelRuns
 from . import plugins