+++ octarine/conversion.py
@@ -1,7 +1,7 @@
 import pygfx as gfx
 import trimesh as tm
 
-from .utils import is_hashable, is_points, is_lines, is_volume, is_mesh_like, is_pygfx_visual, is_pygfx_geometry, VoxelCloud
+from .utils import is_hashable, is_points, is_lines, is_volume, is_mesh_like, is_pygfx_visual, is_pygfx_geometry, VoxelCloud, VoxelRuns
 from .visuals import points2gfx, lines2gfx, mesh2gfx, trimesh2gfx, volume2gfx, scene2gfx, geometry2gfx, visual_passthrough, sparsevolume2gfx
 
 CONVERTERS = {
@@ -10,6 +10,7 @@ CONVERTERS = {
     tm.Trimesh: trimesh2gfx,
     tm.Scene: scene2gfx,
     VoxelCloud: sparsevolume2gfx,
+    VoxelRuns: sparsevolume2gfx,
     is_mesh_like: mesh2gfx,
     is_points: points2gfx,
     is_lines: lines2gfx,